backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.7
Pillow==10.0.1
openai==1.54.4
httpx==0.27.2
spacy==3.7.2
PyPDF2==3.0.1
python-docx==0.8.11
cloudinary==1.36.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0 